Setup ebtables only for mac filtered setups
We used to drop DHCP requests for all bridged setups in case of
nfdhcpd. This requires ebtables setup for private vlans as well.

Dropping DHCP requests in case of nfdhcpd is not really needed
because they are dropped in user space
(payload.set_verdict(nfqueue.NF_DROP)) if a binding file exists.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>