Fix typo in ebtables
If interface has an IP we used to add a rule that does not allow
packets coming from TAP to have a different IP. There was a typo
in if statement and the rule did never apply.

Still, we do NOT want this rule so we fix typo and comment the
ebtables rule.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>