Introduce log and save functions
The save() function is used in order to save and backup
iptables/arptables/ebtables commands related to each interface.

The log() function is a wrapper of snf-network-log script.

Do not use a default variable for the log script. This could
cause a fork bomb if the admin forgot to define it.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>